+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/characters/AbstractPlayer/GiveOrbSlotOnChannel.java
@@ -1,18 +1,19 @@
 package basemod.patches.com.megacrit.cardcrawl.characters.AbstractPlayer;
 
+import basemod.BaseMod;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.orbs.AbstractOrb;
 
 @SpirePatch(
-		cls="com.megacrit.cardcrawl.characters.AbstractPlayer",
+		clz=AbstractPlayer.class,
 		method="channelOrb"
 )
 public class GiveOrbSlotOnChannel
 {
 	public static void Prefix(AbstractPlayer __instance, AbstractOrb orbToSet)
 	{
-		if (__instance.masterMaxOrbs == 0 && __instance.maxOrbs == 0) {
+		if (BaseMod.fixesEnabled && __instance.masterMaxOrbs == 0 && __instance.maxOrbs == 0) {
 			__instance.masterMaxOrbs = 1;
 			__instance.increaseMaxOrbSlots(1, true);
 		}