+++ CHANGELOG.md
@@ -858,3 +858,7 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix crash in ZHT with modded keywords (modargo)
 
 #### dev ####
+* Add option to disable BaseMod gameplay fixes, allowing a more vanilla experience (kiooeht)
+  * Automatically give orb slot
+  * Allow more than 5 combat rewards
+* Add tooltip option to ModLabeledToggleButton (kiooeht)